build: harden CI and CMake for v1.0.0
- Bump version to 1.0.0
- Add Ninja generator to all CI pipelines
- Default COBRA_BUILD_TESTS and COBRA_BUILD_LLVM_PASS to OFF
- Gate GoogleTest behind COBRA_BUILD_TESTS in superbuild
- Fix C++ standard mismatch: superbuild deps now build as C++23
- Fix CMAKE_CXX_FLAGS overwrite with CMAKE_CXX_FLAGS_INIT
- Fix cobra-verify z3 linkage (PRIVATE → PUBLIC) and add PIC
- Add LLVM_FOUND guard in CobraPass CMakeLists
- Add find_dependency for absl/hwy in CobraConfig.cmake.in
- Add -Wall -Wextra -Wpedantic compiler warnings
- Add USE_EXTERNAL_HIGHWAY option for system highway
- Move forwarding configs to cmake/config/ templates
- Split lint into independent CI job
- Add LLVM script SHA256 verification
- Add compiler/tool version logging step
- Hash-based cache key from dependencies/*.cmake
- Parallel ctest with nproc/sysctl fallback
- Update macOS runner to macos-15
- Split dependabot groups (infrastructure vs tools)
- Update BUILD.md prerequisites to C++23
- Update README dependency comments and optional test instructions

+++ lib/core/AnfCleanup.cpp
@@ -256,11 +256,11 @@ namespace cobra {
                 if (kSubsetSize < 2) { continue; }
 
                 uint32_t var_mask = 0;
-                for (int b = 0; b < kNv; ++b) {
+                for (size_t b = 0; b < static_cast< size_t >(kNv); ++b) {
                     if ((s & (1U << b)) != 0u) { var_mask |= var_bits[b]; }
                 }
 
-                const uint32_t kFamilySize = (1ULL << kSubsetSize) - 1;
+                const uint32_t kFamilySize = static_cast< uint32_t >((1ULL << kSubsetSize) - 1);
                 // Skip if this covers all monomials (Rule 1 handles)
                 if (kFamilySize >= monomials.size()) { continue; }
